+++ CHANGELOG.md
@@ -13,6 +13,8 @@ automatically by the auto instrumentor.
 ([#745](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/745))
 - `opentelemetry-instrumentation-pika` Bugfix use properties.headers. It will prevent the header injection from raising.
   ([#740](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/740))
+- `opentelemetry-instrumentation-botocore` Add extension for DynamoDB
+  ([#735](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/735))
 
 ## [1.6.0-0.25b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.6.0-0.25b0) - 2021-10-13
 ### Added